requirements.txt
scapy==2.6.1
requests==2.32.3
python-dotenv==1.1.0
